requirements.txt
asgiref==3.2.10
autopep8==1.5.4
beautifulsoup4==4.9.1
dj-database-url==0.5.0
Django==3.1.1
django-appconf==1.0.4
django-bootstrap-form==3.4
django-bootstrap4==2.2.0
django-bower==5.2.0
django-cleanup==5.0.0
django-comments-dab==1.6.7
django-filer==2.0.2
django-filter==2.3.0
django-js-asset==1.2.2
django-libs==2.0.3
django-mptt==0.11.0
django-polymorphic==3.0.0
django-qrcode==0.3
django-taggit==1.3.0
django-tinymce==3.0.2
django-user-accounts==3.0.2
easy-thumbnails==2.7
gunicorn==20.0.4
Pillow==7.2.0
git+https://github.com/HiveTechies/pinax-templates-change
psycopg2==2.8.6
pycodestyle==2.6.0
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
Unidecode==1.1.1
whitenoise==5.2.0
django-heroku